Don't put current word in SearchView in WordActivity
Instead, add it as a suggestion for an empty query.  With this change,
searching for something based on the current word takes one touch more
(to select the refine query button on the suggestion), while searching
for something completely different takes one touch less (no need to
clear the search box).  In my usage, the later action is more frequent
than the former.