Improve <p> CSS
Instead of applying to zero margins to all <P>, zero only the one
following the headword (so that we don't get a gap at the beginning) and
the 3rd <P> (so that we don't get a gap between the header and the first
definition in the case of words with multiple definitions).  The rest
can have the default margins, which gives us a reasonable amount of
space between different defintions.