Only show WebView after post-processing
Only show the WebView after the Javascript post-processing is done, to
avoid momentarily displaying a blank page.

If we adjust the visibility of the elements inside a RelativeLayout a
resize of the WebView is visible, so use a FrameLayout instead.  The
WebView is rendered in the background and becomes visible when the
LinearLayout is hidden.